moved Config scenario to a file
