+++ tomatic/minizinc.py
@@ -1,96 +1,21 @@
 import asyncio
 from tomato_cooker.grill import GrillTomatoCooker
 from tomato_cooker.models import TomaticProblem, tomatic
-from consolemsg import step, error
-from yamlns import namespace as ns
-import datetime
+from consolemsg import step
 import random
-from .retriever import addDays
-from .shiftload import ShiftLoadComputer
 from .backtracker import parseArgs
-from tomatic.retriever import (
-    downloadPersons,
-    downloadLeaves,
-    downloadIdealLoad,
-    downloadVacations,
-    downloadFestivities,
-    downloadBusy,
-    downloadShiftload,
-    downloadOverload,
-    addDays,
-)
-
-WEEKDAY = {
-    'dl': 0,
-    'dm': 1,
-    'dx': 2,
-    'dj': 3,
-    'dv': 4,
-}
-
-
-# TODO: extract this
-class Config:
-
-    def __init__(self, config_file, date = None, keep = False):
-        step('Carregant configuració {}...', config_file)
-        try:
-            self.data = ns.load(config_file)
-            self._update_monday(date)
-            not keep and downloadPersons(self.data)
-            self._update_persons()
-            if not self.data.get('idealshifts'):
-                self.data.idealshifts = 'idealshifts.yaml'
-                not keep and downloadIdealLoad(self.data, args.certificate)
-            if not self.data.get('weekShifts') and not self.data.computeShifts:
-                self.data.weekShifts = 'carrega.csv'
-                not keep and downloadShiftload(self.data)
-            if not self.data.computeShifts:
-                self.data.overloadfile = "overload-{}.yaml".format(self.data.monday)
-                not keep and downloadOverload(self.data)
-            not keep and self._download_leaves()
-            if not self.data.get('busyFiles'):
-                not keep and self._download_busy()
-            if self.data.computeShifts:
-                self.update_shifts()
-        except:
-            error("Configuració incorrecta")
-            raise
-
-    def update_shifts(self):
-        setup = ShiftLoadComputer.loadData(self.data)
-        self.data.idealLoad = setup.idealLoad
-        self.data.busyTable = setup.busyTable._table
-
-    def _update_persons(self):
-        from .persons import persons
-        self.data.update(persons(self.data.get('personsfile', None)))
-
-    def _update_monday(self, date):
-        if date is not None:
-            # take the monday of the week including that date
-            givenDate = datetime.datetime.strptime(date,"%Y-%m-%d").date()
-            self.data.monday = addDays(givenDate, -givenDate.weekday())
-        else:
-            # If no date provided, take the next monday
-            today = datetime.date.today()
-            self.data.monday = addDays(today, 7-today.weekday())
-
-    def _download_leaves(self):
-        self.data.driveCertificate = args.certificate
-        downloadLeaves(self.data, args.certificate)
-
-    def _download_busy(self):
-        downloadBusy(self.data)
-        downloadFestivities(self.data)
-        downloadVacations(self.data, source=args.holidays)
-
-    def set_ignore_optionals(self, ignore = False):
-        self.data.ignoreOptionalAbsences = ignore
-
+from .scenario_config import Config
 
 
 class Menu:
+
+    WEEKDAY = {
+        'dl': 0,
+        'dm': 1,
+        'dx': 2,
+        'dj': 3,
+        'dv': 4,
+    }
 
     def __init__(self, config):
         self.nPersones = len(config.idealLoad)
@@ -116,7 +41,7 @@ class Menu:
         }
         for day, turn, name in config.busyTable:
             if config.busyTable[(day, turn, name)]:
-                persons_indisponibilities[name][WEEKDAY[day]].add(turn + 1)
+                persons_indisponibilities[name][self.WEEKDAY[day]].add(turn + 1)
 
         indisponibilities = []
         for name in self.names:
@@ -162,7 +87,13 @@ def solve_problem(config, solvers):
 def main():
     global args
     args = parseArgs()
-    config = Config(args.config_file, args.date, args.keep)
+    config = Config(
+        args.config_file,
+        args.date,
+        args.keep,
+        args.certificate,
+        args.holidays
+    )
     # Fist try to get a solution with optional absences
     step('Provant amb les indisponibilitats opcionals...')
     # choose a list of minizinc solvers to user
